Pull request #234: AG-25296 - Instrumented tests improvement
Merge in ADGUARD-CORE-LIBS/vpn-libs from feature/instrumented_test_improvement to master

Squashed commit of the following:

commit 15ef6ca356e06aa79d1a3a0eae5a33fc43adc569
Author: Max Grupper <m.grupper@adguard.com>
Date:   Thu Oct 19 13:33:09 2023 +0300

    skipci: don't close browser while navigation is active

commit 5de468d80b8288a08034ea695bc3e0243ae252c8
Author: Max Grupper <m.grupper@adguard.com>
Date:   Thu Oct 19 12:17:38 2023 +0300

    skipci: add notifications

commit ca63438b1b00d916a0310515990d8ce8e6c517e1
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Oct 18 17:42:38 2023 +0300

    skipci: remove verbose

commit a8a7b686f326ed4910cec00e74c49d29c909716c
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Oct 18 17:34:55 2023 +0300

    skipci: format

commit 1d5cd71cd46dcf96fc8a2fa3ab6a774f1d5cbcd6
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Oct 18 17:14:53 2023 +0300

    increase log size

commit a4d5abe40359dd96a86c7febbcd41b0a993bc5e7
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Oct 18 15:29:40 2023 +0300

    set real limits

commit 608ebaa2d5eb788618e790559f8e53d4f228419c
Merge: 2850dab 56e4272
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Oct 18 15:15:47 2023 +0300

    Merge branch 'master' into feature/instrumented_test_improvement

commit 2850dabb0004cabd451f8d46286ece9de3673f9c
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Oct 18 14:56:26 2023 +0300

    remove flag

commit 3f4af7bb6bb739c5e7416e4da676d4c23636ad58
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Oct 18 14:53:47 2023 +0300

    fix

commit 292657fc3105bd80471d60faf18463f9ca44c2d1
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Oct 18 14:44:21 2023 +0300

    kill each process

commit b893cb5737aa364b094b923633619ce9df233748
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Oct 18 14:36:13 2023 +0300

    test

commit 6eb6b7a3aed48f71072954cc444685b95341ca12
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Oct 18 14:22:45 2023 +0300

    try this

commit d4e6b777f414b2675392c6df72473241e9772e9e
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Oct 18 14:09:03 2023 +0300

    send hup signal

commit 4767b4a5d0acc2a9e76a9ec9ca768f1f0ddf4a7b
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Oct 18 13:45:00 2023 +0300

    show processes

commit aaa046bca9951a2f39856b3781327f157dc40085
Merge: 2fa559d 40ed33c
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Oct 18 13:28:53 2023 +0300

    Merge remote-tracking branch 'origin/feature/instrumented_test_improvement' into feature/instrumented_test_improvement

    # Conflicts:
    #	integration-tests/integration_test.sh

commit 2fa559da4d2206b2f7b7940574a24667a5008404
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Oct 18 11:18:47 2023 +0300

    another way to drop inet

commit 40ed33cfad3e0f286693693ba256146ef648e574
Author: Max Grupper <m.grupper@adguard.com>
Date:   Wed Oct 18 11:18:47 2023 +0300

    another way to drop inet

commit 93bc518953ef7c1d9d9c3c898c6f641edb188752
Author: Max Grupper <m.grupper@adguard.com>
Date:   Tue Oct 17 15:57:03 2023 +0300

    wait longer

commit c275ef2d757ca3ae1efed415b25e177321dc0023
Author: Max Grupper <m.grupper@adguard.com>
Date:   Tue Oct 17 13:41:03 2023 +0300

    ping non-exclusion

commit 9dbbe79acf657b8f0bb196da33d07db86e418da3
Author: Max Grupper <m.grupper@adguard.com>
Date:   Tue Oct 17 12:51:16 2023 +0300

    ping after networkloss

... and 57 more commits

+++ integration-tests/integration_test.sh
@@ -27,19 +27,44 @@ LOG_FILE_NAME="vpn.log"
 
 ENDPOINT_CONTAINER=""
 CLIENT_CONTAINER=""
+COMMON_CONTAINER=""
+
+ADGUARD_API_DOMAIN=""
+ADGUARD_API_CREDS_URL=""
+ADGUARD_API_LOCATIONS_URL=""
 
 BAMBOO_CONAN_REPO_URL=""
+BAMBOO_VPN_APP_ID=""
+BAMBOO_VPN_TOKEN=""
+BAMBOO_ADGUARD_API_CREDS_PATH=""
+BAMBOO_ADGUARD_API_LOCATIONS_PATH=""
+BAMBOO_ADGUARD_API_DOMAIN=""
+BAMBOO_ADGUARD_RELAY_IP=""
+
+NETWORK_SIMULATOR_CONFIG_FILE="network-load-simulator/config.conf"
 
 clean_client() {
-  docker rmi -f "$CLIENT_IMAGE"
+  if docker image inspect "$CLIENT_IMAGE" > /dev/null 2>&1; then
+    docker rmi -f "$CLIENT_IMAGE"
+  fi
 }
 
 clean_client_with_image() {
-  docker rmi -f "$CLIENT_WITH_BROWSER_IMAGE"
+  if docker image inspect "$CLIENT_WITH_BROWSER_IMAGE" > /dev/null 2>&1; then
+    docker rmi -f "$CLIENT_WITH_BROWSER_IMAGE"
+  fi
 }
 
 clean_endpoint() {
-  docker rmi -f "$ENDPOINT_IMAGE"
+  if docker image inspect "$ENDPOINT_IMAGE" > /dev/null 2>&1; then
+    docker rmi -f "$ENDPOINT_IMAGE"
+  fi
+}
+
+clean_common() {
+  if docker image inspect "$COMMON_IMAGE" > /dev/null 2>&1; then
+    docker rmi -f "$COMMON_IMAGE"
+  fi
 }
 
 clean_network() {
@@ -77,6 +102,22 @@ build_all() {
   build_endpoint
 }
 
+build_config_file() {
+  echo -e "ENDPOINT_HOSTNAME=${ENDPOINT_HOSTNAME}" >> ${NETWORK_SIMULATOR_CONFIG_FILE}
+  echo -e "ENDPOINT_IP=${ENDPOINT_IP}" >> ${NETWORK_SIMULATOR_CONFIG_FILE}
+  echo -e "APP_ID=${BAMBOO_VPN_APP_ID}" >> ${NETWORK_SIMULATOR_CONFIG_FILE}
+  echo -e "TOKEN=${BAMBOO_VPN_TOKEN}" >> ${NETWORK_SIMULATOR_CONFIG_FILE}
+  echo -e "CREDS_API_URL=${ADGUARD_API_CREDS_URL}" >> ${NETWORK_SIMULATOR_CONFIG_FILE}
+  echo -e "PROTOCOL=${PROTOCOL}" >> ${NETWORK_SIMULATOR_CONFIG_FILE}
+  echo -e "MODE=${MODE}" >> ${NETWORK_SIMULATOR_CONFIG_FILE}
+  echo -e "LOG_FILE_NAME=${LOG_FILE_NAME}" >> ${NETWORK_SIMULATOR_CONFIG_FILE}
+  echo -e "SOCKS_PORT=${SOCKS_PORT}" >> ${NETWORK_SIMULATOR_CONFIG_FILE}
+}
+
+run_common() {
+  COMMON_CONTAINER=$(docker run --rm -d --entrypoint /bin/bash ${COMMON_IMAGE} -c "while true; do sleep 1000; done")
+}
+
 run_client_tun() {
   PROTOCOL=$1
   LOG_FILE_NAME="vpn_tun_$PROTOCOL.log"
@@ -94,8 +135,6 @@ run_client_tun() {
 }
 
 run_client_with_browser() {
-  PROTOCOL=$1
-  LOG_FILE_NAME="vpn_tun_$PROTOCOL.log"
   CLIENT_WITH_BROWSER_CONTAINER=$(docker run -d --rm \
     -v $SELF_DIR_PATH/logs:/output \
     --cap-add=NET_ADMIN \
@@ -104,8 +143,7 @@ run_client_with_browser() {
     --add-host="$ENDPOINT_HOSTNAME":"$ENDPOINT_IP" \
     --sysctl net.ipv6.conf.all.disable_ipv6=1 \
     --sysctl net.ipv6.conf.default.disable_ipv6=1 \
-    "$CLIENT_WITH_BROWSER_IMAGE" \
-    "$ENDPOINT_HOSTNAME" "$ENDPOINT_IP" "$PROTOCOL" "tun" "$LOG_FILE_NAME")
+    "$CLIENT_WITH_BROWSER_IMAGE" 2>&1)
   echo "Client container with browser run: $CLIENT_WITH_BROWSER_CONTAINER"
 }
 
@@ -139,6 +177,8 @@ stop_containers() {
 }
 
 clean() {
+  rm -f ${NETWORK_SIMULATOR_CONFIG_FILE}
+  clean_common
   clean_network
   clean_client
   clean_endpoint
@@ -159,20 +199,49 @@ run_tun_test() {
   exit "$RESULT"
 }
 
+get_location_data() {
+    docker cp network-load-simulator/get_location.sh ${COMMON_CONTAINER}:get_location.sh
+    docker exec ${COMMON_CONTAINER} /bin/bash -c "bash ./get_location.sh ${ADGUARD_API_LOCATIONS_URL}"
+    docker cp ${COMMON_CONTAINER}:result.txt network-load-simulator/endpoint.txt
+    ENDPOINT_HOSTNAME=$(cat network-load-simulator/endpoint.txt | grep ENDPOINT_HOSTNAME | cut -d '=' -f 2)
+    ENDPOINT_IP=$BAMBOO_ADGUARD_RELAY_IP
+    docker stop ${COMMON_CONTAINER}
+    rm -f network-load-simulator/endpoint.txt
+}
+
 run_browser_test() {
   build_common
+  run_common
+
+  ADGUARD_API_DOMAIN=$BAMBOO_ADGUARD_API_DOMAIN
+  ADGUARD_API_CREDS_URL="https://${ADGUARD_API_DOMAIN}${BAMBOO_ADGUARD_API_CREDS_PATH}"
+  ADGUARD_API_LOCATIONS_URL="https://${ADGUARD_API_DOMAIN}${BAMBOO_ADGUARD_API_LOCATIONS_PATH}"
+  get_location_data
+
+  PROTOCOL=http2
+  LOG_FILE_NAME="vpn_tun_http2.log"
+  MODE="tun"
+  build_config_file
   build_client_with_browser
-  build_endpoint
+
   RESULT=0
-  for protocol in http2; do
-    run_endpoint
-    run_client_with_browser $protocol
-    docker exec -w /test "$CLIENT_WITH_BROWSER_CONTAINER" node index.js || RESULT=1
-    docker exec "$CLIENT_WITH_BROWSER_CONTAINER" chmod -R 777 /output/
-    docker cp $CLIENT_WITH_BROWSER_CONTAINER:/test/output.json ./output.json
-    docker stop "$ENDPOINT_CONTAINER"
-    docker stop "$CLIENT_WITH_BROWSER_CONTAINER"
-  done
+  run_client_with_browser
+
+  sleep 5
+  docker exec -w /test -e TIME_LIMIT=30m "$CLIENT_WITH_BROWSER_CONTAINER" node index.js || RESULT=1
+  docker exec "$CLIENT_WITH_BROWSER_CONTAINER" chmod -R 777 /output/
+  docker cp "$CLIENT_WITH_BROWSER_CONTAINER":/test/output.json ./output1part.json
+
+  docker exec "$CLIENT_WITH_BROWSER_CONTAINER" /bin/bash -c "iptables -A OUTPUT -j DROP; iptables -A INPUT -j DROP"
+  sleep 1
+  docker exec "$CLIENT_WITH_BROWSER_CONTAINER" /bin/bash -c 'pids=$(pgrep standalone); echo "PIDS: $pids"; for pid in $pids; do kill -SIGHUP $pid || true; done'
+  sleep 9
+  docker exec "$CLIENT_WITH_BROWSER_CONTAINER" /bin/bash -c "iptables -D OUTPUT -j DROP; iptables -D INPUT -j DROP"
+  sleep 60
+
+  docker exec -w /test -e TIME_LIMIT=30m "$CLIENT_WITH_BROWSER_CONTAINER" node index.js || RESULT=1
+  docker cp $CLIENT_WITH_BROWSER_CONTAINER:/test/output.json ./output2part.json
+  docker stop "$CLIENT_WITH_BROWSER_CONTAINER"
   exit "$RESULT"
 }
 
@@ -195,6 +264,12 @@ run() {
   elif [[ "$MODE" == "socks" ]]; then
     run_socks_test
   elif [[ "$MODE" == "browser" ]]; then
+    BAMBOO_VPN_APP_ID=$1
+    BAMBOO_VPN_TOKEN=$2
+    BAMBOO_ADGUARD_API_CREDS_PATH=$3
+    BAMBOO_ADGUARD_API_LOCATIONS_PATH=$4
+    BAMBOO_ADGUARD_API_DOMAIN=$5
+    BAMBOO_ADGUARD_RELAY_IP=$6
     run_browser_test
   fi
 }
@@ -203,7 +278,8 @@ WORK=$1
 if [[ "$WORK" == "run" ]]; then
   MODE=$2
   BAMBOO_CONAN_REPO_URL=$3
-  run
+  shift 3
+  run "$@"
 elif [[ "$WORK" == "clean-browser" ]]; then
   clean_client_with_image
 elif [[ "$WORK" == "clean" ]]; then
